backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.0.2
scikit-learn==1.5.2
pydantic==2.8.2
mistralai>=1.0.0
python-multipart==0.0.9